COM-2016 refacto maintenance modal
